health-analysis-service/requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
pytesseract
Pillow
pdf2image
torch
torchvision
torchxrayvision
gtts
numpy
scikit-learn
xgboost
transformers
tensorflow
